cookbook/sdxl_prompt_optimize/requirement.txt
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
numpy
Pillow
platformdirs
requests
torch
torchvision
tqdm
accelerate @ git+https://github.com/huggingface/accelerate.git@d1aa558119859c4b205a324afabaecabd9ef375e
datasets==2.10.1
deepspeed==0.8.3
fire==0.4.0
hydra-core==1.3.2
rich==13.3.2
submitit==1.4.5
transformers==4.27.3
wandb==0.12.21